SHA-1: * 1. Target : TPTab 2. Reason    - TP Connection 전체 메시지 출력기능 업데이트 3. Solution    - DataGridView를 이용하여, DB 혹은 Configuration 파일에서 읽어온 TP Connection 전에 정보 출려기능 업데이트

+++ MenuFrame/FileMenu/FileMenuOperation.cs
@@ -753,7 +753,7 @@ namespace bsw_generation.MenuFrame.FileMenu
                                     break;
 
                                 case "PAD":
-                                    tpInformation.Pad = Convert.ToUInt16(loadConfigInformation.Value);
+                                    tpInformation.Pad = Convert.ToByte(loadConfigInformation.Value);
                                     break;
 
                                 default:
